SolaraAIQuantDashboard/BE/requirements.txt
Django>=4.2,<5.2
django-cors-headers>=4.0
djangorestframework>=3.14
psycopg2-binary>=2.9
python-dotenv>=1.0
celery>=5.3
redis>=5.0
django-celery-beat>=2.5
django-celery-results>=2.5
